You can now set the images for backgrounds of badges. only have attendee on the settings page, but what is done could easily be duplicated
I don't think the image helper class is being used anymore. the PngResult class is a simplified class that just responds with a png and proper content type. It is used for serving the images for previewing.

the getimage method in settings controller ignores it's parameter and always responds with one image

the routes could also be updated to make that look nicer

refs #91