working on being able to send volunteer time slots
updated database so time slots know about the integers for scheduling (not used) and that the link can contain a comment (possibly for who is being recorded)

added attribute for string contains. is much nicer that huge regular expressions. TODO: remove the large regular expressions that check for it and use [StringContains("foo")] instead

references #3